+++ src/N98/Magento/Command/Developer/ConsoleCommand.php
@@ -5,6 +5,7 @@ namespace N98\Magento\Command\Developer;
 use Exception;
 use Magento\Framework\App\ProductMetadataInterface;
 use N98\Magento\Command\AbstractMagentoCommand;
+use N98\Magento\Command\Developer\Console\Shell;
 use N98\Util\Unicode\Charset;
 use PhpParser\Lexer;
 use PhpParser\Parser;
@@ -13,7 +14,8 @@ use Psy\Command\ListCommand;
 use Psy\Configuration;
 use Psy\Output\ShellOutput;
 use Psy\ParserFactory;
-use Psy\Shell;
+use Psy\TabCompletion\AutoCompleter;
+use Psy\TabCompletion\Matcher\CommandsMatcher;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -58,14 +60,30 @@ class ConsoleCommand extends AbstractMagentoCommand
             // do nothing
         }
 
-        $parser = new Parser(new Lexer());
-        $cleaner = new CodeCleaner($parser);
-        $consoleOutput = new ShellOutput();
         $config = new Configuration();
+
+        $php5Parser = new Parser\Php5(new Lexer\Emulative());
+        $php7Parser = new Parser\Php7(new Lexer\Emulative());
+
+        $parser = new Parser\Multiple([$php5Parser, $php7Parser]);
+        $cleaner = new CodeCleaner($parser);
         $config->setCodeCleaner($cleaner);
+
+        $consoleOutput = new ShellOutput();
+
+        $commandConfig = $this->getCommandConfig();
+        $commandsToAdd = [];
+        foreach ($commandConfig['commands'] as $command) {
+            $commandsToAdd[]= new $command();
+        }
+
+        $config->addCommands($commandsToAdd);
+
         $shell = new Shell($config);
         $shell->setScopeVariables([
             'di' => $this->getObjectManager(),
+            'magerun' => $this->getApplication(),
+            'magerunInternal' => (object)['currentModule' => ''],
         ]);
 
         if ($initialized) {