+++ tests/N98/Magento/Command/Developer/Console/MakeHelperCommandTest.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace N98\Magento\Command\Developer\Console;
+
+
+use Magento\Framework\Filesystem\Directory\WriteInterface;
+
+class MakeHelperCommandTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function testOutput()
+    {
+        $command = new MakeHelperCommand();
+
+        $commandTester = $this->createCommandTester($command);
+        $command->setCurrentModuleName('N98_Dummy');
+
+        $writerMock = $this->getMock(WriteInterface::class);
+        $writerMock
+            ->expects($this->once())
+            ->method('writeFile')
+            ->with(
+                $this->anything(), // param1
+                $this->equalTo(file_get_contents(__DIR__ . '/_files/reference_helper.php'))
+            );
+
+        $command->setCurrentModuleDirectoryWriter($writerMock);
+        $commandTester->execute(['classpath' => 'foo.bar.baz']);
+    }
+}