+++ tests/N98/Magento/Command/Developer/Console/_files/reference_block.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace N98\Dummy\Block\Foo\Bar;
+
+use Magento\Framework\View\Element\Template;
+
+class Baz extends Template
+{
+
+
+}
+