+++ tests/N98/Magento/Command/Developer/Console/_files/reference_interface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace N98\Dummy\Foo\Bar;
+
+interface Baz
+{
+
+
+}
+