+++ tests/N98/Magento/Command/Developer/Console/_files/reference_model.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace N98\Dummy\Model\Foo\Bar;
+
+use Magento\Framework\Model\AbstractModel;
+
+class Baz extends AbstractModel
+{
+
+
+}
+